src/requirements.txt
python-dateutil
requests
pyyaml
fastjsonschema